+++ routes/feedback.mjs
@@ -2,7 +2,7 @@ import Router from 'koa-router';
 import _get from 'lodash.get';
 import Luxon from 'luxon';
 import config from '../services/config';
-import checkRights from '../middlewares/check-rights';
+import resolveRights from '../middlewares/check-rights';
 import sendMail from '../services/mail';
 import { CAN_SEND_FEEDBACK } from '../models/rights';
 
@@ -11,7 +11,7 @@ const router = new Router();
 
 router.post(
   '/',
-  checkRights(CAN_SEND_FEEDBACK),
+  resolveRights(CAN_SEND_FEEDBACK),
   async (ctx) => {
     const {
       request: { body: { message, type } },