+++ routes/index.mjs
@@ -12,8 +12,14 @@ import jwt from './jwt';
 import campus from './campuses';
 import feedback from './feedback';
 import forms from './forms';
+import phones from './phones';
+import phoneModels from './phone-models';
 
 const router = new Router();
+
+router.get('/', (ctx) => {
+  ctx.body = 'OK';
+});
 
 router.use('/users', users);
 router.use('/pois', pois);
@@ -28,5 +34,7 @@ router.use('/jwt', jwt);
 router.use('/campuses', campus);
 router.use('/feedback', feedback);
 router.use('/forms', forms);
+router.use('/phones', phones);
+router.use('/phone-models', phoneModels);
 
 export default router.routes();