+++ tests/routes/campuses.test.mjs
@@ -1,50 +1,55 @@
-import chai from 'chai';
-import request, { generateSuperAdminJWTHeader, generateRegulatorJWTHeader } from '../request';
+import {
+  generateSuperAdminJWTHeader,
+  generateRegulatorJWTHeader,
+  generateDriverJWTHeader,
+  generateAnonymousJWTHeader,
+} from '../request';
 import Campus, { generateDummyCampus } from '../models/campus';
+import {
+  testCreate, testCreateUnicity, testDelete, testList, testGet, testUpdate,
+} from '../helpers/crud';
 
-const { expect } = chai;
+const config = {
+  route: '/campuses',
+  generateDummyObject: generateDummyCampus,
+};
 
-describe('Test the campus route', () => {
-  it('POST API endpoint should return an error when the campus already exists', async () => {
-    const dummyCampus = new Campus(generateDummyCampus());
-    try {
-      {
-        const { body: campus, statusCode } = await request()
-          .post('/campuses')
-          .set(...generateSuperAdminJWTHeader())
-          .query({ mask: 'id,name' })
-          .send({
-            id: dummyCampus._id,
-            name: dummyCampus.name,
-          });
-        expect(statusCode).to.equal(200);
-        expect(campus.id).to.equal(dummyCampus._id);
-        expect(campus.name).to.equal(dummyCampus.name);
-        const campusFound = await Campus
-          .findById(dummyCampus._id)
-          .lean();
-        expect(campusFound).to.not.be.null;
-      }
-      {
-        const { statusCode } = await request()
-          .post('/campuses')
-          .set(...generateSuperAdminJWTHeader())
-          .send({
-            id: dummyCampus._id,
-            name: dummyCampus.name,
-          });
-        expect(statusCode).to.equal(409);
-      }
-    } finally {
-      await dummyCampus.remove();
-    }
-  });
+describe('Test the campuses route', () => {
+  it(...testCreate(Campus, {
+    ...config,
+    cannotCall: [generateRegulatorJWTHeader, generateDriverJWTHeader],
+    canCall: [generateSuperAdminJWTHeader],
+  }));
 
-  it('It should response the GET method', async () => {
-    const { body: list, statusCode } = await request()
-      .get('/campuses')
-      .set(...generateRegulatorJWTHeader());
-    expect(statusCode).to.equal(200);
-    expect(Array.isArray(list)).to.be.true;
-  });
+  it(...testCreateUnicity(Campus, {
+    ...config,
+    requestCallBack: r => r
+      .set(...generateSuperAdminJWTHeader()),
+    transformObject: { id: '_id', name: 'name' },
+  }));
+
+  it(...testList(Campus, {
+    ...config,
+    canCall: [generateAnonymousJWTHeader],
+  }));
+
+  it(...testDelete(Campus, {
+    ...config,
+    route: ({ id }) => `${config.route}/${id}`,
+    cannotCall: [generateRegulatorJWTHeader, generateDriverJWTHeader],
+    canCall: [generateSuperAdminJWTHeader],
+  }));
+
+  it(...testGet(Campus, {
+    ...config,
+    route: ({ id }) => `${config.route}/${id}`,
+    canCall: [generateDriverJWTHeader],
+  }));
+
+  it(...testUpdate(Campus, {
+    ...config,
+    route: ({ id }) => `${config.route}/${id}`,
+    cannotCall: [generateRegulatorJWTHeader, generateDriverJWTHeader],
+    canCall: [generateSuperAdminJWTHeader],
+  }));
 });